# internal/docs/swaggerdocs/swagger.yaml
definitions:
  github_com_linkeunid_go-api_internal_model.Animal:
    properties:
      age:
        example: 3
        maximum: 200
        minimum: 0
        type: integer
      created_at:
        type: string
      description:
        example: A friendly cat with white fur
        maxLength: 1000
        type: string
      id:
        type: integer
      name:
        example: Fluffy
        maxLength: 100
        minLength: 2
        type: string
      species:
        example: Cat
        maxLength: 100
        minLength: 2
        type: string
      updated_at:
        type: string
    required:
    - name
    - species
    type: object
  github_com_linkeunid_go-api_internal_model.AnimalCreateRequest:
    properties:
      age:
        example: 3
        type: integer
      description:
        example: A friendly cat with white fur
        type: string
      name:
        example: Fluffy
        type: string
      species:
        example: Cat
        type: string
    type: object
  github_com_linkeunid_go-api_internal_model.AnimalUpdateRequest:
    properties:
      age:
        example: 3
        type: integer
      description:
        example: A friendly cat with white fur
        type: string
      name:
        example: Fluffy
        type: string
      species:
        example: Cat
        type: string
    type: object
  github_com_linkeunid_go-api_pkg_pagination.PagedData:
    properties:
      cacheInfo: {}
      items: {}
      pagination:
        $ref: '#/definitions/github_com_linkeunid_go-api_pkg_pagination.Params'
    type: object
  github_com_linkeunid_go-api_pkg_pagination.Params:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_linkeunid_go-api_pkg_response.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
info:
  contact: {}
paths:
  /animals:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all animals
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Sort field (id, name, species, age, created_at, updated_at)
        in: query
        name: sort
        type: string
      - description: Sort direction (asc, desc)
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_linkeunid_go-api_pkg_pagination.PagedData'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_linkeunid_go-api_internal_model.Animal'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
      summary: Get all animals
      tags:
      - animals
    post:
      consumes:
      - application/json
      description: Create a new animal with the provided details
      parameters:
      - description: Animal object to be created
        in: body
        name: animal
        required: true
        schema:
          $ref: '#/definitions/github_com_linkeunid_go-api_internal_model.AnimalCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_linkeunid_go-api_internal_model.Animal'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
      summary: Create a new animal
      tags:
      - animals
  /animals/{animalID}:
    delete:
      consumes:
      - application/json
      description: Delete an animal by its ID
      parameters:
      - description: Animal ID
        in: path
        name: animalID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
      summary: Delete an animal
      tags:
      - animals
    get:
      consumes:
      - application/json
      description: Get an animal by its ID
      parameters:
      - description: Animal ID
        in: path
        name: animalID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_linkeunid_go-api_internal_model.Animal'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
      summary: Get an animal by ID
      tags:
      - animals
    put:
      consumes:
      - application/json
      description: Update an existing animal by its ID
      parameters:
      - description: Animal ID
        in: path
        name: animalID
        required: true
        type: string
      - description: Updated animal object
        in: body
        name: animal
        required: true
        schema:
          $ref: '#/definitions/github_com_linkeunid_go-api_internal_model.AnimalUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_linkeunid_go-api_internal_model.Animal'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_linkeunid_go-api_pkg_response.APIResponse'
      summary: Update an animal
      tags:
      - animals
swagger: "2.0"